Add storage server RPC client (#1430)
* Add storage server RPC client

* Address review comments

- Use `#[tracing::instrument]` wherever appropriate
- Remove `Mutex` for `client` and `clone` it wherever required
- Avoid reading env variable inside the lib, rather take a config object
- Make the private methods of `DurableWal` async
- Don't try to create a runtime, instead assume it always exists. Let the caller create it
- Update proto:
	- remove `max_frame_no` from `InsertFramesRequest`
	- make `page_no` to `u32` in the proto
- Update storage server to have `page_no` as `u32`

+++ libsql-storage-server/src/store.rs
@@ -3,18 +3,18 @@ use bytes::Bytes;
 
 #[async_trait]
 pub trait FrameStore: Send + Sync {
-    async fn insert_frame(&self, namespace: &str, page_no: u64, frame: bytes::Bytes) -> u64;
+    async fn insert_frame(&self, namespace: &str, page_no: u32, frame: bytes::Bytes) -> u64;
     #[allow(dead_code)]
     async fn insert_frames(&self, namespace: &str, frames: Vec<FrameData>) -> u64;
     async fn read_frame(&self, namespace: &str, frame_no: u64) -> Option<bytes::Bytes>;
-    async fn find_frame(&self, namespace: &str, page_no: u64) -> Option<u64>;
-    async fn frame_page_no(&self, namespace: &str, frame_no: u64) -> Option<u64>;
+    async fn find_frame(&self, namespace: &str, page_no: u32) -> Option<u64>;
+    async fn frame_page_no(&self, namespace: &str, frame_no: u64) -> Option<u32>;
     async fn frames_in_wal(&self, namespace: &str) -> u64;
     async fn destroy(&self, namespace: &str);
 }
 
 #[derive(Default)]
 pub struct FrameData {
-    pub(crate) page_no: u64,
+    pub(crate) page_no: u32,
     pub(crate) data: Bytes,
 }
